+++ src/content_delegate.js
@@ -622,42 +622,6 @@ ContentDelegate.prototype.handleSingleDocumentPageView = function () {
     'message', this.onDocumentViewSubmit.bind(this), false);
 };
 
-// Pop up a dialog offering the link to the free cached copy of the document,
-// or just go directly to the free document if popups are turned off.
-ContentDelegate.prototype.handleRecapLinkClick = function (window_obj, url) {
-  chrome.storage.local.get('options', function (items) {
-    if (!items.options.recap_link_popups) {
-      window_obj.location = url;
-      return;
-    }
-    $('<div id="recap-shade"/>').appendTo($('body'));
-    $('<div class="recap-popup"/>').append(
-      $('<a/>', {
-        'class': 'recap-close-link',
-        href: '#',
-        onclick: 'var d = document; d.body.removeChild(this.parentNode); ' +
-          'd.body.removeChild(d.getElementById("recap-shade")); return false'
-      }).append(
-        '\u00d7'
-      )
-    ).append(
-      $('<a/>', {
-        href: url,
-        onclick: 'var d = document; d.body.removeChild(this.parentNode); ' +
-          'd.body.removeChild(d.getElementById("recap-shade"))'
-      }).append(
-        ' Get this document for free from RECAP.'
-      )
-    ).append(
-      $('<br><br><small>Note that archived documents may be out of date. ' +
-        'RECAP is not affiliated with the U.S. Courts. The documents ' +
-        'it makes available are voluntarily uploaded by PACER users. ' +
-        'RECAP cannot guarantee the authenticity of documents because the ' +
-        'courts provide no effective document authentication system.</small>')
-    ).appendTo($('body'));
-  });
-  return false;
-};
 
 // Check every link in the document to see if there is a free RECAP document
 // available. If there is, put a link with a RECAP icon.
@@ -684,13 +648,12 @@ ContentDelegate.prototype.attachRecapLinkToEligibleDocs = function () {
         if (!result) {
           continue;
         }
-        let href = `https://www.courtlistener.com/${result.filepath_local}`;
+        let href = `https://storage.courtlistener.com/${result.filepath_local}`;
         let recap_link = $('<a/>', {
           'class': 'recap-inline',
           'title': 'Available for free from the RECAP Archive.',
           'href': href
         });
-        recap_link.click($.proxy(this.handleRecapLinkClick, this, window, href));
         recap_link.append($('<img/>').attr({
           src: chrome.extension.getURL('assets/images/icon-16.png')
         }));